using Col and moving people finally
